feat: sig status constraint for leaving the sig

+++ src/routes/pig.py
@@ -118,6 +118,7 @@ async def join_pig(id: int, session: SessionDep, request: Request):
     if not pig: raise HTTPException(404, detail="해당 id의 시그/피그가 없습니다")
     allowed = (ctrl_status_available.join_sigpig_rolling_admission if pig.is_rolling_admission else ctrl_status_available.join_sigpig)
     if pig.status not in allowed: raise HTTPException(400, f"시그/피그 상태가 {allowed}일 때만 시그/피그에 가입할 수 있습니다")
+
     pig_member = PIGMember(
         ig_id=id,
         user_id=current_user.id,
@@ -139,7 +140,10 @@ async def leave_pig(id: int, session: SessionDep, scsc_global_status: SCSCGlobal
     current_user = get_user(request)
     pig = session.get(PIG, id)
     if not pig: raise HTTPException(404, detail="해당 id의 시그/피그가 없습니다")
+    allowed = (ctrl_status_available.join_sigpig_rolling_admission if pig.is_rolling_admission else ctrl_status_available.join_sigpig)
+    if pig.status not in allowed: raise HTTPException(400, f"시그/피그 상태가 {allowed}일 때만 시그/피그에서 탈퇴할 수 있습니다")
     if pig.owner == current_user.id: raise HTTPException(409, detail="시그/피그장은 해당 시그/피그를 탈퇴할 수 없습니다")
+
     pig_member = session.exec(select(PIGMember).where(PIGMember.ig_id == id).where(PIGMember.user_id == current_user.id).where(PIGMember.status == scsc_global_status.status)).first()
     if not pig_member: raise HTTPException(404, detail="시그/피그의 구성원이 아닙니다")
     session.delete(pig_member)
